Update relative paths in multiple files
This commit revises the relative paths used in various source files and build configuration files. The paths have been shortened by one level, presuming that the overall directory hierarchy of the project has been slightly restructured. Also, there was a minor adjustment in the linked libraries in the CMakeLists file.