Update lua script and CMakeLists.txt for FFI integration
The lua script ('main.lua') has been updated to demonstrate usage of the FFI library. Furthermore, changes in 'CMakeLists.txt' involve updating the C++ standard to C++17 and the transformation of the project from a shared library into an executable. Also added FFI library linking based on the operating system.