Update function execution and increase CMake CXX standard
In `raia_llvm.cpp`, the method for executing the function has been updated. Instead of using the execution engine directly, the function address is now obtained and cast to a function pointer type followed by function execution. The CMake CXX standard has been updated to 17 in the `CMakeLists.txt` file.